refactoring: separate sign in with connection channel
